restructure App
